Split options for request into a different file
