+++ src/models/block.js
@@ -1,18 +1,10 @@
 var request = require("request-promise");
 var async = require("async");
+var Options = require("../utils/options");
 
 module.exports = {
   getBlockByHash: async hash => {
-    var options = {
-      url: process.env.URL,
-      method: "POST",
-      body: JSON.stringify({
-        jsonrpc: "1.0",
-        id: "curltest",
-        method: "getblock",
-        params: [hash]
-      })
-    };
+    var options = Options("getblock", [hash]);
 
     var data = await request(options);
     data = JSON.parse(data).result
@@ -20,33 +12,23 @@ module.exports = {
     return data;
   },
   getBlockByHeight: async (height) => {
-    var options = {
-      url: process.env.URL,
-      method: "POST",
-      body: JSON.stringify({
-        jsonrpc: "1.0",
-        id: "curltest",
-        method: "getblockhash",
-        params: [height]
-      })
-    };
+    var options = Options("getblockhash", [height]);
 
     var data = await request(options);
     var hash = JSON.parse(data).result;
 
-    var options = {
-      url: process.env.URL,
-      method: "POST",
-      body: JSON.stringify({
-        jsonrpc: "1.0",
-        id: "curltest",
-        method: "getblock",
-        params: [hash]
-      })
-    };
+    options = Options("getblock", [hash]);
 
     data = await request(options);
-    data = JSON.parse(data).result
+    data = JSON.parse(data).result;
+
+    return data;
+  },
+  getBestBlockHash: async () => {
+    var options = Options("getbestblockhash", []);
+
+    var data = await request(options);
+    data = JSON.parse(data).result;
 
     return data;
   }
